@@ -1,18 +1,21 @@
 {
   "name": "vite",
+  "private": true,
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build"
   },
   "dependencies": {
-    "vant": "^4.0.0",
-    "vue": "^3.2.45"
+    "vant": "^4.8.5",
+    "vue": "^3.4.21"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^3.2.0",
+    "@vant/auto-import-resolver": "^1.1.0",
+    "@vitejs/plugin-vue": "^4.3.4",
     "less": "^4.1.3",
-    "unplugin-vue-components": "^0.22.11",
-    "vite": "^3.2.4"
+    "unplugin-auto-import": "^0.17.5",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.1.6"
   }
 }
